refactor: separate concerns for snapshot manager
