+++ libs/python/agent/agent/callbacks/snapshot_manager/__init__.py
@@ -1,0 +1,7 @@
+"""
+Snapshot manager callback components.
+"""
+
+from .callback import SnapshotManagerCallback
+
+__all__ = ["SnapshotManagerCallback"]
